rename package to singular
