+++ whatsforlunch/src/main/java/com/richmond/whatsforlunch/service/UserService.java
@@ -1,8 +1,8 @@
-package com.richmond.whatsforlunch.services;
+package com.richmond.whatsforlunch.service;
 
 import com.richmond.whatsforlunch.respository.UserRepository;
 import com.richmond.whatsforlunch.respository.entity.UserEntity;
-import com.richmond.whatsforlunch.services.dto.User;
+import com.richmond.whatsforlunch.service.dto.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
