+++ whatsforlunch/src/main/java/com/richmond/whatsforlunch/service/dto/User.java
@@ -1,4 +1,4 @@
-package com.richmond.whatsforlunch.services.dto;
+package com.richmond.whatsforlunch.service.dto;
 
 public record User(long id, String userName, String firstName, String lastName) {
 }
