+++ whatsforlunch/src/test/java/com/richmond/whatsforlunch/controller/UserControllerTest.java
@@ -1,7 +1,7 @@
-package com.richmond.whatsforlunch.controllers;
+package com.richmond.whatsforlunch.controller;
 
-import com.richmond.whatsforlunch.services.UserService;
-import com.richmond.whatsforlunch.services.dto.User;
+import com.richmond.whatsforlunch.service.UserService;
+import com.richmond.whatsforlunch.service.dto.User;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
